move RefIdItem to common
